requirements.txt
asgiref==3.2.10
autopep8==1.5.4
Django==3.1.2
mysqlclient==2.0.1
Pillow==8.0.0
python-decouple==3.3